# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FixMyMind API",
        "description": "API for therapeutic voice and text chat with AI",
        "contact": {
            "email": "contact@fixmymind.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.fixmymind.org/api",
            "description": "Production API Server"
        },
        {
            "url": "/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/therapist-preferences": {
            "get": {
                "tags": [
                    "Therapist Preferences"
                ],
                "summary": "Get user's therapist preferences",
                "operationId": "26737beb680875f4c5fa9470b67d8928",
                "responses": {
                    "200": {
                        "description": "Therapist preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Alex"
                                                },
                                                "style": {
                                                    "type": "string",
                                                    "example": "warm"
                                                },
                                                "tempo": {
                                                    "type": "string",
                                                    "example": "slow"
                                                },
                                                "tone": {
                                                    "type": "string",
                                                    "example": "soft"
                                                },
                                                "add_markers": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "ask_questions": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "generated_prompt": {
                                                    "type": "string",
                                                    "example": "Jesteś terapeutą o imieniu Alex..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No preferences found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No therapist preferences found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Therapist Preferences"
                ],
                "summary": "Create or update therapist preferences",
                "operationId": "f547cc7b735e29499eeb1b516f21c49d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Therapist name",
                                        "type": "string",
                                        "example": "Alex"
                                    },
                                    "style": {
                                        "type": "string",
                                        "enum": [
                                            "warm",
                                            "calm",
                                            "analytical",
                                            "motivating",
                                            "deep"
                                        ],
                                        "example": "warm"
                                    },
                                    "tempo": {
                                        "type": "string",
                                        "enum": [
                                            "slow",
                                            "medium",
                                            "fast"
                                        ],
                                        "example": "slow"
                                    },
                                    "tone": {
                                        "type": "string",
                                        "enum": [
                                            "soft",
                                            "neutral",
                                            "firm"
                                        ],
                                        "example": "soft"
                                    },
                                    "add_markers": {
                                        "description": "Add pauses and expressions like 'I understand', 'hmm'",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "ask_questions": {
                                        "description": "Ask reflective questions",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Therapist preferences saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Therapist preferences saved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Alex"
                                                },
                                                "style": {
                                                    "type": "string",
                                                    "example": "warm"
                                                },
                                                "tempo": {
                                                    "type": "string",
                                                    "example": "slow"
                                                },
                                                "tone": {
                                                    "type": "string",
                                                    "example": "soft"
                                                },
                                                "add_markers": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "ask_questions": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "generated_prompt": {
                                                    "type": "string",
                                                    "example": "Jesteś terapeutą o imieniu Alex..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/therapist-preferences/options": {
            "get": {
                "tags": [
                    "Therapist Preferences"
                ],
                "summary": "Get available options for therapist preferences",
                "operationId": "773dcc5f5b928a26e4da4ce9f4e4ff3b",
                "responses": {
                    "200": {
                        "description": "Available options retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "styles": {
                                                    "properties": {
                                                        "warm": {
                                                            "type": "string",
                                                            "example": "Ciepły i wspierający"
                                                        },
                                                        "calm": {
                                                            "type": "string",
                                                            "example": "Spokojny i refleksyjny"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "tempos": {
                                                    "properties": {
                                                        "slow": {
                                                            "type": "string",
                                                            "example": "Wolno"
                                                        },
                                                        "medium": {
                                                            "type": "string",
                                                            "example": "Średnio"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "tones": {
                                                    "properties": {
                                                        "soft": {
                                                            "type": "string",
                                                            "example": "Miękki"
                                                        },
                                                        "neutral": {
                                                            "type": "string",
                                                            "example": "Neutralny"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}